% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qi_FUN.R
\name{logistic_prob_FUN}
\alias{logistic_prob_FUN}
\title{Function for finding predicted probability from the systematic linear
component of a logistic regression.}
\usage{
logistic_prob_FUN(x)
}
\arguments{
\item{x}{a numeric vector of the systematic linear component from a
logistic regression model.}
}
\description{
Function for finding predicted probability from the systematic linear
component of a logistic regression.
}
