% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.plte.R
\name{plte_plot}
\alias{plte_plot}
\title{Plot objects created by plte_builder}
\usage{
plte_plot(obj, ...)
}
\arguments{
\item{obj}{a \code{plte} class object created by \code{\link{plte_builder}}.}
}
\description{
Plot objects created by plte_builder
}

