% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qi_builder}
\alias{qi_builder}
\title{Find quantities of interest from generalized linear models}
\usage{
qi_builder(b_sims, newdata, model, ...)
}
\arguments{
\item{b_sims}{a data frame created by \code{\link{b_sim}} of simulated
coefficients}

\item{newdata}{a data frame of fitted values with column names corresponding
to variable names in \code{b_sims}. If \code{missing} then a normal
linear regression model is assumed and the predicted values are returned
(e.g. the fitted linear systematic component from
\code{\link{linear_systematic}}).}

\item{model}{a function for calculating how}

\item{...}{arguments to pass to \code{\link{linear_systematic}}}
}
\description{
Find quantities of interest from generalized linear models
}

